# new object pool to remove boost dependency