@mrgreywater 2016-02-13
see #28