@mrgreywater 2016-02-13
Code before was something like that: 
```cpp
    template <typename T>
    class BlockPool {
    public:
        BlockPool() : blockSize(0) { }
        BlockPool(std::size_t blockSize) {
            reset(blockSize);
        }
        template <typename... Args>
        T* construct(Args&&... args) {
            if (block->size() == blockSize) {
                assert(blockSize);
                blocks.emplace_back();
                block++;
                block->reserve(blockSize);
            }
            block->emplace_back(std::forward<Args>(args)...);
            return (T*)(&block->operator[](block->size()-1));
        }
        void reset(std::size_t new_size) {
            blockSize = new_size;
            blocks.clear();
            blocks.emplace_back();
            block = std::prev(blocks.end());
            block->reserve(blockSize);
        }
    private:
        typedef std::list<std::vector<T>> block_t;
        std::size_t blockSize = 0;
        block_t blocks;
        typename block_t::iterator block;
    };
```
But sadly it didn't perform as well.