@mrgreywater 2016-02-13
On possible addition would be to reuse allocated blocks for later polygon triangulations, but considering the api will change in a way that this is not possible anymore (see #17), it would unnecessarily increase the size of the class.